[IF-319] Add CI/CD pipeline
This pipeline builds provider for every OS arch, creates release, and uploads the assets

- Matrix strategy is used to build all binary versions in parallel
- Generate metadata file for TF Registry
- Generate SHA256 file for all zips
- Sign SHA256 file with GPG key

This process is fully compatible with the TF Registry process. If we decide to make it public it should work out of the box. We only need to configure the TF registry, upload the GPG key, and setup webhook.